arduino: cast to long before substracting unsigned integers
The result of substracting two unsigned integers is always > 0! Thus,
cast the first unsigned integer to long before doing any substraction.

This should fix a bug where the reset message did not work if the
current position was higher than the initial position.

Found by clang.

+++ src/plugins/arduino/com_thread.cpp
@@ -150,13 +150,13 @@ ArduinoComThread::loop()
             } else if (arduino_if_->msgq_first_is<ArduinoInterface::ResetZPosMessage>()) {
                 if((cfg_init_mm_ - current_z_position_) != 0) {
                     ArduinoComMessage req;
-                    if ((cfg_init_mm_ - current_z_position_) < 0) {
+                    if ((long) cfg_init_mm_ - current_z_position_ < 0) {
                         req.set_command(ArduinoComMessage::CMD_STEP_UP);
                     } else {
                         req.set_command(ArduinoComMessage::CMD_STEP_DOWN);
                     }
-                    req.set_number(abs(cfg_init_mm_ - current_z_position_) * ArduinoComMessage::NUM_STEPS_PER_MM);
-                    req.set_msecs(((double) abs((cfg_init_mm_ - current_z_position_) * ArduinoComMessage::NUM_STEPS_PER_MM) / (double)cfg_speed_) * 1000. * 10.);
+                    req.set_number(abs((long) cfg_init_mm_ - current_z_position_) * ArduinoComMessage::NUM_STEPS_PER_MM);
+                    req.set_msecs(((double) abs(((long) cfg_init_mm_ - current_z_position_) * ArduinoComMessage::NUM_STEPS_PER_MM) / (double)cfg_speed_) * 1000. * 10.);
                     logger->log_debug(name(), "Moving to init_mm position, %u steps", (cfg_init_mm_ - current_z_position_) * ArduinoComMessage::NUM_STEPS_PER_MM);
                     messages_.push(req);
                 }
